feat: src/pages/AlbumDetailPageのコンポーネントを分割する
