--- src/components/SongListTable.stories.tsx
@@ -1,20 +1,0 @@
-/* eslint-disable import/no-default-export */
-
-import { ComponentMeta, ComponentStoryObj } from "@storybook/react";
-import { ComponentProps } from "react";
-import { Component } from "./SongListTable";
-import { mockSongs } from "../mocks/mockData";
-
-export default {
-  title: "components/SongListTable",
-  component: Component,
-} as ComponentMeta<typeof Component>;
-
-type Props = ComponentProps<typeof Component>;
-
-export const SongListTable: ComponentStoryObj<typeof Component> = {
-  args: {
-    songs: mockSongs,
-    handleClickSong: () => void 0,
-  } as Props,
-};
