+++ src/components/model/albums/AlbumDisplay/index.ts
@@ -1,0 +1,1 @@
+export { AlbumDisplay } from "./AlbumDisplay";
