+++ src/components/model/albums/AlbumListItem/stories.tsx
