+++ src/components/model/songs/SongListTable/index.ts
@@ -1,0 +1,1 @@
+export { SongListTable } from "./SongListTable";
