+++ src/components/model/songs/SongListTableRow/index.ts
@@ -1,0 +1,1 @@
+export { SongListTableRow } from "./SongListTableRow";
