+++ src/components/model/songs/SongListTableRow/styles.tsx
@@ -1,0 +1,28 @@
+import styled from "styled-components";
+
+import { TableCell, TableRow, tableCellClasses } from "@mui/material";
+import { styled as muiStyled } from "@mui/material/styles";
+
+export const StyledTableCell = muiStyled(TableCell)(({ theme }) => ({
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: theme.palette.common.black,
+    color: theme.palette.common.white,
+  },
+  [`&.${tableCellClasses.body}`]: {
+    fontSize: 14,
+  },
+}));
+
+export const StyledTableRow = muiStyled(TableRow)(({ theme }) => ({
+  "&:nth-of-type(odd)": {
+    backgroundColor: theme.palette.action.hover,
+  },
+  "&:last-child td, &:last-child th": {
+    border: 0,
+  },
+}));
+
+export const StyledAnchor = styled.a({
+  color: "#444",
+  cursor: "pointer",
+});
