+++ src/pages/AlbumDetailPage.tsx
@@ -1,14 +1,9 @@
 import { Link, useLocation } from "react-router-dom";
-import {
-  StyledHeading,
-  StyledImage,
-  StyledParagraph,
-  StyledContainer,
-} from "./styles";
-import { ServiceList } from "../components/ServiceList";
-import { SongListTable } from "../components/SongListTable";
-import { useFetch } from "../hooks/useFetch";
+import { StyledHeading } from "./styles";
 import { fetchAlbum, fetchSongs } from "../lib";
+import { useFetch } from "../hooks/useFetch";
+import { SongListTable } from "../components/model/songs/SongListTable";
+import { AlbumDisplay } from "../components/model/albums/AlbumDisplay";
 
 export const AlbumDetailPage: React.FC = () => {
   const location = useLocation();
@@ -23,17 +18,8 @@ export const AlbumDetailPage: React.FC = () => {
   return (
     <article className="space-bottom">
       <StyledHeading>{album?.title}</StyledHeading>
-      <StyledContainer>
-        <StyledImage src={album?.imageFile.path} alt={album?.title} />
-        {album?.publishPlatform && (
-          <ServiceList services={album?.publishPlatform} />
-        )}
-        {album?.description !== "" && (
-          <StyledParagraph>{album?.description}</StyledParagraph>
-        )}
-      </StyledContainer>
-
-      {songs && songs.length > 0 && <SongListTable songs={songs} />}
+      {album && <AlbumDisplay album={album} />}
+      {songs && <SongListTable songs={songs} />}
       <br />
       <Link to="/">もどる</Link>
     </article>
